+++ src/main/groovy/io/seqera/nf/diff/DagComparator.groovy
@@ -23,8 +23,6 @@ import java.nio.file.Paths
 import java.nio.file.SimpleFileVisitor
 import java.nio.file.attribute.BasicFileAttributes
 
-import java.nio.file.Path
-
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import io.seqera.nf.diff.DiffResult.DagEdge